Split default_env
